Add `DynamicModule`
