+++ packages/core/src/decorators/EncryptedModule.ts
@@ -1,9 +1,9 @@
 import { IEncryptionPassword } from "@nestia/fetcher/lib/IEncryptionPassword";
 import { Module, ModuleMetadata } from "@nestjs/common";
-import is_ts_node from "detect-ts-node";
-import fs from "fs";
 
+import { Creator } from "../typings/Creator";
 import { ENCRYPTION_METADATA_KEY } from "./internal/EncryptedConstant";
+import { load_controllers } from "./internal/load_controller";
 
 /**
  * Encrypted module.
@@ -43,16 +43,9 @@ export function EncryptedModule(
         Module(metadata)(target);
         if (metadata.controllers === undefined) return;
 
-        for (const controller of metadata.controllers)
-            if (
-                Reflect.hasMetadata(ENCRYPTION_METADATA_KEY, controller) ===
-                false
-            )
-                Reflect.defineMetadata(
-                    ENCRYPTION_METADATA_KEY,
-                    password,
-                    controller,
-                );
+        for (const c of metadata.controllers)
+            if (Reflect.hasMetadata(ENCRYPTION_METADATA_KEY, c) === false)
+                Reflect.defineMetadata(ENCRYPTION_METADATA_KEY, password, c);
     };
 }
 
@@ -74,54 +67,11 @@ export namespace EncryptedModule {
         password: IEncryptionPassword | IEncryptionPassword.Closure,
     ): Promise<object> {
         // LOAD CONTROLLERS
-        const metadata: ModuleMetadata = {
-            controllers: await controllers(path, password),
-        };
+        const controllers: Creator<object>[] = await load_controllers(path);
 
         // RETURNS WITH DECORATING
-        @EncryptedModule(metadata, password)
-        class Module {}
-        return Module;
-    }
-
-    async function controllers(
-        path: string,
-        password: IEncryptionPassword | IEncryptionPassword.Closure,
-    ): Promise<any[]> {
-        const output: any[] = [];
-        await iterate(output, path);
-
-        for (const controller of output)
-            if (
-                Reflect.hasMetadata(ENCRYPTION_METADATA_KEY, controller) ===
-                false
-            )
-                Reflect.defineMetadata(
-                    ENCRYPTION_METADATA_KEY,
-                    password,
-                    controller,
-                );
-
-        return output;
-    }
-
-    async function iterate(controllers: object[], path: string): Promise<void> {
-        const directory: string[] = await fs.promises.readdir(path);
-        for (const file of directory) {
-            const current: string = `${path}/${file}`;
-            const stats: fs.Stats = await fs.promises.lstat(current);
-
-            if (stats.isDirectory() === true)
-                await iterate(controllers, current);
-            else if (file.substring(file.length - 3) === `.${EXTENSION}`) {
-                const external: any = await import(current);
-                for (const key in external) {
-                    const instance: object = external[key];
-                    if (Reflect.getMetadata("path", instance) !== undefined)
-                        controllers.push(instance);
-                }
-            }
-        }
+        @EncryptedModule({ controllers }, password)
+        class NestiaModule {}
+        return NestiaModule;
     }
 }
-const EXTENSION = is_ts_node ? "ts" : "js";
