+++ packages/core/src/decorators/internal/route_error.ts
@@ -4,9 +4,6 @@ import { Observable, throwError } from "rxjs";
 
 import { ExceptionManager } from "../../utils/ExceptionManager";
 
-/**
- * @internal
- */
 export function route_error(
     request: express.Request,
     error: any,
