+++ packages/core/src/module.ts
@@ -1,3 +1,4 @@
+export * from "./decorators/DynamicModule";
 export * from "./decorators/EncryptedBody";
 export * from "./decorators/EncryptedController";
 export * from "./decorators/EncryptedModule";
@@ -7,3 +8,4 @@ export * from "./decorators/PlainBody";
 export * from "./decorators/TypedBody";
 export * from "./decorators/TypedParam";
 export * from "./decorators/TypedRoute";
+export * from "./options/INestiaTransformOptions";
